feat: also log programs to S3
To get rid of speed limits of jsonbin (and to stop having to pay for
it), also log parse log files to S3.

Factor out the `LogQueue` class and reuse it for this logging.

JSON Bin logging has not been turned off yet: it's safer to log to
both destinations in parallel until we're convinced that everythings
work the way we want it to.

Add a script called `tools/download-logs` which can be used to download
all programs from the S3 bucket (and add the AWS CLI dependency to
`requirements.txt`).

**⚠️ Do not merge this PR yet ⚠️**

* #468 should be merged first, even before you start reviewing this:
  this PR will show a lot of unrelated changes until that PR is merged
  (since it has been based off of the new file locations introduced
  in that PR).
* #460 should be merged as well, and after #460 is merged the
  "emergency shutdown" mechanism should be introduced into this PR
  as well, taking into account that there now might be multiple
  queues (so the queue name should be worked into the emergency
  save files).
