+++ config.py
@@ -27,4 +27,11 @@ config = {
         'postfix': ('-' + dyno if dyno else '') + '-' + str(os.getpid()),
         'region': 'eu-west-1'
     },
+    's3-parse-logs': {
+        'bucket': 'hedy-parse-logs',
+        'prefix': app_name + '/',
+        # Make logs from different instances/processes unique
+        'postfix': ('-' + dyno if dyno else '') + '-' + str(os.getpid()),
+        'region': 'eu-west-1'
+    },
 }
