+++ requirements.txt
@@ -7,6 +7,7 @@ requests==2.23.0
 attrs==19.3.0
 Flask-Commonmark==0.8
 bcrypt==3.2.0
-boto3==1.16.50
+boto3>=1.16.50
 ruamel.yaml==0.17.4
 pylint==2.8.2
+awscli>=1.19.88
