+++ website/jsonbin.py
@@ -5,6 +5,9 @@ import queue
 import requests
 import logging
 
+from . import log_queue
+from . import aws_helpers
+
 logger = logging.getLogger('jsonbin')
 
 class JsonBinLogger:
@@ -69,7 +72,41 @@ class JsonBinLogger:
             except Exception:
                 logger.exception(f'Error posting to jsonbin.')
 
+
 class NullJsonbinLogger():
     """A jsonbin logger that doesn't actually do anything."""
     def log(self, obj):
         pass
+
+
+class MultiParseLogger():
+    """A logger that forwards to other loggers."""
+    def __init__(self, *loggers):
+        self.loggers = loggers
+
+    def log(self, obj):
+        for logger in self.loggers:
+            logger.log(obj)
+
+
+class S3ParseLogger():
+    """A logger that logs to S3.
+
+    - Well then why is it in a file called 'jsonbin.py'?
+
+    - Legacy, young grasshopper. Legacy.
+    """
+    @staticmethod
+    def from_env_vars():
+        transmitter = aws_helpers.s3_parselog_transmitter_from_env()
+        if not transmitter:
+            return NullJsonbinLogger()
+
+        S3_LOG_QUEUE.set_transmitter(transmitter)
+        return S3ParseLogger()
+
+    def log(self, obj):
+        S3_LOG_QUEUE.add(obj)
+
+
+S3_LOG_QUEUE = log_queue.LogQueue('parse', batch_window_s=300)
